Almost done... added transforms for solids, and clip triangle which i hope is implemented well :D

+++ src/render/TriangleRasterizer.java
@@ -11,7 +11,7 @@ public class TriangleRasterizer {
     }
 
     public void rasterize(Vertex a, Vertex b, Vertex c) {
-        // Seřazení vrcholů podle y hodnoty
+        // Serazeni vrcholu podle y hodnoty
         Vec2D pa = new Vec2D(a.getPosition().getX(), a.getPosition().getY());
         Vec2D pb = new Vec2D(b.getPosition().getX(), b.getPosition().getY());
         Vec2D pc = new Vec2D(c.getPosition().getX(), c.getPosition().getY());
@@ -29,15 +29,14 @@ public class TriangleRasterizer {
             Vertex tempV = b; b = c; c = tempV;
         }
 
-        // První polovina trojúhelníku (A -> B)
-        for (double y =  pa.getY(); y <=  pb.getY(); y++) {
+        // 1st polovina trojuhelniku
+        for (double y = pa.getY(); y <= pb.getY(); y++) {
             double t1 = (y - pa.getY()) / (pb.getY() - pa.getY());
             double t2 = (y - pa.getY()) / (pc.getY() - pa.getY());
 
             double x1 = (1 - t1) * pa.getX() + t1 * pb.getX();
             double x2 = (1 - t2) * pa.getX() + t2 * pc.getX();
 
-            // Interpolace pro barvy a z-souřadnice
             Vertex ab = a.mul(1 - t1).add(b.mul(t1));
             Vertex ac = a.mul(1 - t2).add(c.mul(t2));
 
@@ -49,19 +48,18 @@ public class TriangleRasterizer {
             for (int x = (int) x1; x <= (int) x2; x++) {
                 double t3 = (x - x1) / (x2 - x1);
                 Vertex abc = ab.mul(1 - t3).add(ac.mul(t3));
-                buffer.drawPixelZTest((int)x, (int)y, abc.getPosition().getZ(), abc.getCol());
+                buffer.drawPixelZTest(x, (int)y, abc.getPosition().getZ(), abc.getCol());
             }
         }
 
-        // Druhá polovina trojúhelníku (B -> C)
-        for (double y =  pb.getY(); y <=  pc.getY(); y++) {
+        // 2nd polovina trojuhelniku
+        for (double y = pb.getY(); y <= pc.getY(); y++) {
             double t1 = (y - pb.getY()) / (pc.getY() - pb.getY());
             double t2 = (y - pa.getY()) / (pc.getY() - pa.getY());
 
             double x1 = (1 - t1) * pb.getX() + t1 * pc.getX();
             double x2 = (1 - t2) * pa.getX() + t2 * pc.getX();
 
-            // Interpolace pro barvy a z-souřadnice
             Vertex bc = b.mul(1 - t1).add(c.mul(t1));
             Vertex ac = a.mul(1 - t2).add(c.mul(t2));
 
@@ -73,7 +71,7 @@ public class TriangleRasterizer {
             for (int x = (int) x1; x <= (int) x2; x++) {
                 double t3 = (x - x1) / (x2 - x1);
                 Vertex abc = bc.mul(1 - t3).add(ac.mul(t3));
-                buffer.drawPixelZTest((int)x, (int)y, abc.getPosition().getZ(), abc.getCol());
+                buffer.drawPixelZTest(x, (int)y, abc.getPosition().getZ(), abc.getCol());
             }
         }
     }
